+++ packages/rut-dom/package.json
@@ -18,7 +18,12 @@
   "engines": {
     "node": ">=8.9.0"
   },
+  "peerDependencies": {
+    "react": "^16.9.0"
+  },
   "dependencies": {
+    "@types/react-dom": "^16.9.3",
+    "react-dom": "^16.11.0",
     "rut": "*"
   }
 }
