+++ packages/rut-dom/src/index.ts
@@ -6,6 +6,8 @@
 import { createAdapter } from 'rut/lib/adapters';
 import DomElement from './DomElement';
 
-export const { render, renderAndWait } = createAdapter(instance => new DomElement(instance));
-
 export * from 'rut';
+export * from './mocks';
+export * from './types';
+
+export const { render, renderAndWait } = createAdapter(instance => new DomElement(instance));
