+++ packages/rut-dom/src/mocks.ts
@@ -1,6 +1,28 @@
+/* eslint-disable complexity */
+
+import React from 'react';
+import { InferHostElementFromEvent } from 'rut';
+import { BaseEvent, SyntheticEvent } from 'rut/lib/adapters';
+import { EventType, EventOptions } from './types';
+
+/**
+ * In both the `mockEvent` and `mockSyntheticEvent` functions below,
+ * you'll notice that every return is being type ignored.
+ * We are doing this so that the generic type takes precendence,
+ * as we want inferrence to be predominantly used.
+ *
+ * For example, if a test is dispatching `findOne('button').dispatch('onClick`)`,
+ * the first argument is typed as `React.MouseEvent<HTMLButtonElement, MouseEvent>`.
+ * We can take advantage of type inferrence by mocking the argument at
+ * the call site, like so: `findOne('button').dispatch('onClick', mockSyntheticEvent('onClick'))`.
+ *
+ * With this pattern, mocks are easily typed, and the underlying event object
+ * structure is close enough for most, if not all of test cases.
+ */
+
 // https://developer.mozilla.org/en-US/docs/Web/Events
 // istanbul ignore next
-export function createEvent(type: string): Event {
+function createEvent(type: string): Event {
   switch (type) {
     case 'animationcancel':
     case 'animationiteration':
@@ -102,4 +124,62 @@ export function createEvent(type: string): Event {
     default:
       return new Event(type);
   }
+}
+
+/**
+ * Mock a DOM `Event` based on type.
+ */
+export function mockEvent<T = Event>(type: string, options?: EventOptions<Element, T>): T {
+  const { currentTarget, target, ...props } = options || {};
+  let event: Event;
+
+  // JSDOM environment does not exist, which means we do not have events.
+  // Return a very custom low-level event object for the time being.
+  if (typeof window === 'undefined') {
+    // @ts-ignore Ignore legacy fields
+    event = new BaseEvent(type);
+  } else {
+    event = createEvent(type);
+  }
+
+  if (target) {
+    Object.defineProperty(event, 'target', {
+      enumerable: true,
+      value: target,
+    });
+  }
+
+  if (currentTarget || target) {
+    Object.defineProperty(event, 'currentTarget', {
+      enumerable: true,
+      value: currentTarget || target,
+    });
+  }
+
+  Object.entries(props).forEach(([prop, value]) => {
+    Object.defineProperty(event, prop, {
+      enumerable: true,
+      value,
+    });
+  });
+
+  // @ts-ignore Infer generics
+  return event;
+}
+
+/**
+ * Mock a React `SyntheticEvent` based on type.
+ */
+export function mockSyntheticEvent<T = React.SyntheticEvent>(
+  type: EventType,
+  options?: EventOptions<InferHostElementFromEvent<T>, T>,
+): T {
+  let eventType = type.toLowerCase();
+
+  if (eventType.startsWith('on')) {
+    eventType = eventType.slice(2);
+  }
+
+  // @ts-ignore Infer generics
+  return new SyntheticEvent(eventType, mockEvent(eventType, options));
 }
