+++ packages/rut/package.json
@@ -35,8 +35,6 @@
   },
   "devDependencies": {
     "@types/prop-types": "^15.7.3",
-    "@types/react-dom": "^16.9.3",
-    "prop-types": "^15.7.2",
-    "react-dom": "^16.11.0"
+    "prop-types": "^15.7.2"
   }
 }
