+++ packages/rut/src/BaseEvent.ts
@@ -1,0 +1,40 @@
+export default class BaseEvent<T = unknown> {
+  bubbles: boolean = true;
+
+  cancelable: boolean = true;
+
+  currentTarget!: T;
+
+  defaultPrevented: boolean = false;
+
+  eventPhase: number = 0;
+
+  isTrusted: boolean = true;
+
+  propagationStopped: boolean = false;
+
+  srcElement: null = null;
+
+  target!: T;
+
+  timeStamp: number;
+
+  type: string;
+
+  constructor(type: string) {
+    this.timeStamp = Date.now();
+    this.type = type;
+  }
+
+  preventDefault(): void {
+    this.defaultPrevented = true;
+  }
+
+  stopImmediatePropagation(): void {
+    this.propagationStopped = true;
+  }
+
+  stopPropagation(): void {
+    this.propagationStopped = true;
+  }
+}
