+++ packages/rut/src/adapters.ts
@@ -1,9 +1,14 @@
 import React from 'react';
+import Element from './Element';
+import BaseEvent from './BaseEvent';
+import SyntheticEvent from './SyntheticEvent';
 import AsyncResult from './AsyncResult';
 import SyncResult from './SyncResult';
 import { globalOptions } from './internals/config';
 import { doAct, doAsyncAct } from './internals/act';
 import { AdapterRendererOptions, RendererOptions } from './types';
+
+export { Element, BaseEvent, SyntheticEvent };
 
 export function createAdapter(createElement: AdapterRendererOptions['createElement']) {
   function render<Props extends object = {}>(
