--- packages/rut/src/helpers/index.ts
@@ -1,2 +1,0 @@
-export * from './predicates';
-export * from './utils';
