+++ packages/rut/src/matchers/toBeElementType.ts
@@ -1,7 +1,7 @@
 import Element from '../Element';
-import { formatValue } from '../helpers';
-import { MatchResult } from '../types';
+import { formatValue } from '../internals/react';
 import { isRutElement } from '../internals/utils';
+import { MatchResult } from '../types';
 
 /**
  * Check that an element is a valid React element type.
