+++ packages/rut/src/mocks/index.ts
@@ -1,2 +1,1 @@
-export * from './event';
 export * from './fetch';
